+++ src/Components/Transformer/RecurringOrderTransformerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BestitKlarnaOrderManagement\Components\Transformer;
+
+use BestitKlarnaOrderManagement\Components\Api\Model\RecurringOrder;
+
+interface RecurringOrderTransformerInterface
+{
+    public function toKlarnaOrder(array $basketData, array $userData, string $currency, $locale): RecurringOrder;
+}