+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to `klarnaordermanagement` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
+### Added
+- Moved services from Klarna Payments to order management
+- Added service classes for recurring orders and customer token
 ### Fixed
 - Restored compatibility for subshops